api/requirements.txt
# MySQL (database)
mysql-connector-python==9.2.0
# Flask (API)
flask==3.1.0
